Implement section visibility util
