+++ superset-frontend/packages/superset-ui-chart-controls/src/sections/advancedAnalytics.tsx
@@ -17,15 +17,13 @@
  * under the License.
  */
 import React from 'react';
-import {
-  t,
-  RollingType,
-  ComparisionType,
-  GenericDataType,
-  ensureIsArray,
-} from '@superset-ui/core';
+import { t, RollingType, ComparisionType } from '@superset-ui/core';
 import { ControlPanelSectionConfig } from '../types';
-import { formatSelectOptions } from '../utils';
+import {
+  formatSelectOptions,
+  isSectionVisible,
+  VisibilityRuleType,
+} from '../utils';
 
 export const advancedAnalyticsControls: ControlPanelSectionConfig = {
   label: t('Advanced analytics'),
@@ -201,20 +199,6 @@ export const advancedAnalyticsControls: ControlPanelSectionConfig = {
       },
     ],
   ],
-  visibility: ({ exploreState }) => {
-    if (exploreState?.datasource && exploreState?.form_data) {
-      const { datasource, form_data } = exploreState;
-
-      if (form_data?.x_axis) {
-        const column = ensureIsArray(datasource.columns).find(
-          (col: { column_name: string }) =>
-            col?.column_name === form_data.x_axis,
-        );
-        if (column?.type_generic !== GenericDataType.TEMPORAL) {
-          return false;
-        }
-      }
-    }
-    return true;
-  },
+  visibility: ({ exploreState }) =>
+    isSectionVisible(VisibilityRuleType.XTEMPORAL, exploreState),
 };
